refactor(ptv3): move BackboneEngine to its own file
Relocate BackboneEngine (and BackboneOutputView) from ptv3_trt.{hpp,cpp} into
backbone_engine.{hpp,cpp}. Pure move plus the include and CMake source entry.
